+++ libraries/Runtime/BridJ/src/main/cpp/bridj/PrimDefs_undef.h
@@ -58,29 +58,29 @@
 
 #ifndef REORDER_VALUE_BYTES_jshort
 #define REORDER_VALUE_BYTES_jshort(peer, lowerIndex, upperIndex) \
-	((((jshort)((unsigned char*)peer)[upperIndex]) << 8) | ((unsigned char*)peer)[lowerIndex])
+	((((jshort)((unsigned char*)(size_t)peer)[upperIndex]) << 8) | ((unsigned char*)peer)[lowerIndex])
 #endif
 
 #ifndef REORDER_VALUE_BYTES_jint
 #define REORDER_VALUE_BYTES_jint(peer, idx0, idx1, idx2, idx3) \
 	( \
-		(((jint)((unsigned char*)peer)[idx3]) << 24) | \
-		(((jint)((unsigned char*)peer)[idx2]) << 16) | \
-		(((jint)((unsigned char*)peer)[idx1]) << 8) | \
-				((unsigned char*)peer)[idx0] \
+		(((jint)((unsigned char*)(size_t)peer)[idx3]) << 24) | \
+		(((jint)((unsigned char*)(size_t)peer)[idx2]) << 16) | \
+		(((jint)((unsigned char*)(size_t)peer)[idx1]) << 8) | \
+				((unsigned char*)(size_t)peer)[idx0] \
 	)
 #endif
 
 #ifndef REORDER_VALUE_BYTES_jlong
 #define REORDER_VALUE_BYTES_jlong(peer, idx0, idx1, idx2, idx3, idx4, idx5, idx6, idx7) \
 	( \
-		(((jlong)((unsigned char*)peer)[idx7]) << 56) | \
-		(((jlong)((unsigned char*)peer)[idx6]) << 48) | \
-		(((jlong)((unsigned char*)peer)[idx5]) << 40) | \
-		(((jlong)((unsigned char*)peer)[idx4]) << 32) | \
-		(((jlong)((unsigned char*)peer)[idx3]) << 24) | \
-		(((jlong)((unsigned char*)peer)[idx2]) << 16) | \
-		(((jlong)((unsigned char*)peer)[idx1]) << 8) | \
-				 ((unsigned char*)peer)[idx0] \
+		(((jlong)((unsigned char*)(size_t)peer)[idx7]) << 56) | \
+		(((jlong)((unsigned char*)(size_t)peer)[idx6]) << 48) | \
+		(((jlong)((unsigned char*)(size_t)peer)[idx5]) << 40) | \
+		(((jlong)((unsigned char*)(size_t)peer)[idx4]) << 32) | \
+		(((jlong)((unsigned char*)(size_t)peer)[idx3]) << 24) | \
+		(((jlong)((unsigned char*)(size_t)peer)[idx2]) << 16) | \
+		(((jlong)((unsigned char*)(size_t)peer)[idx1]) << 8) | \
+				 ((unsigned char*)(size_t)peer)[idx0] \
 	)
 #endif
